pyproject.toml
[tool.poetry]
name = "kanidm_operator"
version = "1.1.0"
description = "Kubernetes Operator to manage Kanidm resources"
authors = ["sbordeyne <simon.bordeyne@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "kanidm_operator"}]

[tool.poetry.dependencies]
python = "^3.11"
kubernetes = "^28.1.0"
tomli = "^2.0.1"
kopf = "^1.36.2"
jinja2 = "^3.1.2"
ruamel-yaml = "^0.18.5"

[tool.poetry.dependencies.kanidm]
git = "git@github.com:yaleman/kanidm.git"
branch = "yaleman/issue282"
subdirectory = "pykanidm"

[tool.poetry.group.dev.dependencies]
pylama = "^8.4.1"
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
